fix: router and header button error
