+++ src/__tests__/App_test1.test.tsx
@@ -3,6 +3,7 @@ import MainPage from "../MainPage";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "../store/store";
+import Wrapper from "../Wrapper";
 
 describe("App tests render", () => {
   test("should render the title", () => {
@@ -10,6 +11,7 @@ describe("App tests render", () => {
       <Provider store={store}>
         <Router>
           <MainPage />
+          <Wrapper />
         </Router>
       </Provider>,
     );
